refactor and use `createContractForOrderLine`
