+++ backend/de.metas.contracts/src/main/java/de/metas/contracts/callorder/CallOrderContractService.java
@@ -22,9 +22,7 @@
 
 package de.metas.contracts.callorder;
 
-import de.metas.bpartner.BPartnerContactId;
 import de.metas.bpartner.BPartnerId;
-import de.metas.bpartner.BPartnerLocationAndCaptureId;
 import de.metas.bpartner.service.IBPartnerDAO;
 import de.metas.contracts.ConditionsId;
 import de.metas.contracts.FlatrateTermId;
@@ -32,9 +30,7 @@ import de.metas.contracts.IFlatrateBL;
 import de.metas.contracts.IFlatrateDAO;
 import de.metas.contracts.flatrate.TypeConditions;
 import de.metas.contracts.impl.FlatrateTermOverlapCriteria;
-import de.metas.contracts.location.adapter.ContractDocumentLocationAdapterFactory;
 import de.metas.contracts.model.I_C_Flatrate_Conditions;
-import de.metas.contracts.model.I_C_Flatrate_Data;
 import de.metas.contracts.model.I_C_Flatrate_Term;
 import de.metas.contracts.model.X_C_Flatrate_Term;
 import de.metas.document.engine.IDocumentBL;
@@ -53,7 +49,6 @@ import de.metas.util.Check;
 import de.metas.util.Services;
 import lombok.NonNull;
 import org.adempiere.exceptions.AdempiereException;
-import org.adempiere.model.InterfaceWrapperHelper;
 import org.compiere.model.I_C_BPartner;
 import org.compiere.model.I_C_Invoice;
 import org.compiere.model.I_C_InvoiceLine;
@@ -65,8 +60,6 @@ import org.compiere.model.I_M_Product;
 import org.springframework.stereotype.Service;
 
 import java.util.Objects;
-
-import static org.adempiere.model.InterfaceWrapperHelper.save;
 
 @Service
 public class CallOrderContractService
@@ -92,51 +85,7 @@ public class CallOrderContractService
 	@NonNull
 	public I_C_Flatrate_Term createCallOrderContract(@NonNull final I_C_OrderLine orderLine)
 	{
-		final I_C_Order order = orderBL.getById(OrderId.ofRepoId(orderLine.getC_Order_ID()));
-
-		final I_C_Flatrate_Term newTerm = InterfaceWrapperHelper.newInstance(I_C_Flatrate_Term.class, orderLine);
-
-		//order references
-		newTerm.setC_OrderLine_Term_ID(orderLine.getC_OrderLine_ID());
-		newTerm.setC_Order_Term_ID(orderLine.getC_Order_ID());
-
-		Check.assume(orderLine.getC_Flatrate_Conditions_ID() > 0, "C_Flatrate_Conditions_ID must be set!");
-		newTerm.setC_Flatrate_Conditions_ID(orderLine.getC_Flatrate_Conditions_ID());
-
-		//qty of the whole contract
-		newTerm.setPlannedQtyPerUnit(orderLine.getQtyEnteredInPriceUOM());
-		newTerm.setC_UOM_ID(orderLine.getPrice_UOM_ID());
-
-		//price & tax
-		newTerm.setM_Product_ID(orderLine.getM_Product_ID());
-		newTerm.setPriceActual(orderLine.getPriceActual());
-		newTerm.setC_Currency_ID(orderLine.getC_Currency_ID());
-
-		Check.assume(orderLine.getC_TaxCategory_ID() > 0, "C_TaxCategory_ID must be set!");
-		newTerm.setC_TaxCategory_ID(orderLine.getC_TaxCategory_ID());
-		newTerm.setIsTaxIncluded(order.isTaxIncluded());
-
-		//contract dates
-		newTerm.setStartDate(order.getDatePromised());
-		newTerm.setMasterStartDate(order.getDatePromised());
-
-		//bill partner info
-		final BPartnerLocationAndCaptureId billToLocationId = orderBL.getBillToLocationId(order);
-
-		final BPartnerContactId billToContactId = BPartnerContactId.ofRepoIdOrNull(billToLocationId.getBpartnerId(), order.getBill_User_ID());
-		ContractDocumentLocationAdapterFactory
-				.billLocationAdapter(newTerm)
-				.setFrom(billToLocationId, billToContactId);
-
-		final I_C_BPartner billPartner = bPartnerDAO.getById(billToLocationId.getBpartnerId());
-		final I_C_Flatrate_Data existingData = flatrateDAO.retrieveOrCreateFlatrateData(billPartner);
-		newTerm.setC_Flatrate_Data(existingData);
-
-		//doc
-		newTerm.setDocStatus(X_C_Flatrate_Term.DOCSTATUS_Drafted);
-		newTerm.setDocAction(X_C_Flatrate_Term.DOCACTION_Complete);
-
-		save(newTerm);
+		final I_C_Flatrate_Term newTerm = flatrateBL.createContractForOrderLine(orderLine);
 
 		documentBL.processEx(newTerm, X_C_Flatrate_Term.DOCACTION_Complete, X_C_Flatrate_Term.DOCSTATUS_Completed);
